feerate: use u32 everywhere.
The wire protocol uses this, in the assumption that we'll never see feerates
in excess of 4294967 satoshi per kiloweight.

So let's use that consistently internally as well.

+++ common/initial_channel.h
@@ -20,7 +20,7 @@ struct fulfilled_htlc;
 /* View from each side */
 struct channel_view {
 	/* Current feerate in satoshis per 1000 weight. */
-	u64 feerate_per_kw;
+	u32 feerate_per_kw;
 
 	/* How much is owed to each side (includes pending changes) */
 	u64 owed_msat[NUM_SIDES];
